+++ bfg9000/builtins/compile.py
@@ -458,7 +458,8 @@ try:
 
     @msbuild.rule_handler(GenerateSource)
     def msbuild_generate_source(rule, build_inputs, solution, env):
-        raise ValueError('msbuild backend does not currently support ' +
-                         "'generated_source'")  # pragma: no cover
+        raise ValueError(
+            "msbuild backend does not currently support 'generated_source'"
+        )  # pragma: no cover
 except ImportError:  # pragma: no cover
     pass