+++ bfg9000/builtins/copy_file.py
@@ -1,13 +1,13 @@
 import os
 
 from . import builtin
-from .file_types import FileList, make_file_list
+from .file_types import FileList, make_file_list, static_file
 from .path import buildpath, relname, within_directory
 from ..backends.compdb import writer as compdb
 from ..backends.make import writer as make
 from ..backends.ninja import writer as ninja
 from ..build_inputs import Edge
-from ..file_types import File
+from ..file_types import File, ManPage
 from ..iterutils import first, unlistify
 from ..path import Path
 from ..versioning import SpecifierSet
@@ -46,6 +46,29 @@ class CopyFile(Edge):
         return output, file, kwargs
 
 
+# XXX: It might be worth generalizing this and merging it with CopyFile above,
+# e.g. if we want to allow users to easily compress any kind of file in their
+# build scripts.
+class CompressFile(Edge):
+    mode = 'gzip'
+
+    def __init__(self, context, output, file, *, extra_deps=None,
+                 description=None):
+        self.copier = context.env.tool('gzip')
+        self.file = file
+        super().__init__(context.build, output, None, extra_deps, description)
+
+    @staticmethod
+    def convert_args(context, file, kwargs):
+        file = context['auto_file'](file)
+
+        def pathfn(file):
+            return file.path.reroot().addext('.gz')
+
+        output = file.clone(pathfn)
+        return output, file, kwargs
+
+
 @builtin.function()
 @builtin.type(File, short_circuit=False, first_optional=True)
 def copy_file(context, name, file, **kwargs):
@@ -63,7 +86,30 @@ def copy_files(context, files, **kwargs):
     return make_file_list(context, context['copy_file'], files, **kwargs)
 
 
-@make.rule_handler(CopyFile)
+@builtin.function()
+@builtin.type(ManPage, short_circuit=False)
+def man_page(context, name, *, compress='auto', **kwargs):
+    if isinstance(name, ManPage):
+        file = name
+    else:
+        path = context['relpath'](name)
+        level = kwargs.pop('level', None)
+        if level is None:
+            ext = path.ext()
+            if not ext[1].isdigit():
+                raise ValueError('unable to guess man page level')
+            level = ext[1]
+        dist = kwargs.pop('dist', True)
+        file = static_file(context, ManPage, name, dist, [('level', level)])
+
+    if ( not compress or
+         (compress == 'auto' and not context.env.tool('gzip').found) ):
+        return file
+    output, file, kwargs = CompressFile.convert_args(context, file, kwargs)
+    return CompressFile(context, output, file, **kwargs).public_output
+
+
+@make.rule_handler(CopyFile, CompressFile)
 def make_copy_file(rule, build_inputs, buildfile, env):
     copier = rule.copier
     recipename = make.var('RULE_{}'.format(copier.rule_name.upper()))
@@ -86,7 +132,7 @@ def make_copy_file(rule, build_inputs, buildfile, env):
     )
 
 
-@ninja.rule_handler(CopyFile)
+@ninja.rule_handler(CopyFile, CompressFile)
 def ninja_copy_file(rule, build_inputs, buildfile, env):
     copier = rule.copier
 
@@ -133,6 +179,8 @@ def compdb_copy_file(rule, build_inputs, buildfile, env):
 
 
 try:
+    import warnings
+
     from ..backends.msbuild import writer as msbuild
 
     @msbuild.rule_handler(CopyFile)
@@ -160,5 +208,11 @@ try:
             dependencies=solution.dependencies(rule.extra_deps),
         )
         solution[rule.output[0]] = project
+
+    @msbuild.rule_handler(CompressFile)
+    def msbuild_generate_source(rule, build_inputs, solution, env):
+        warnings.warn(
+            'msbuild backend does not currently support compressing files'
+        )  # pragma: no cover
 except ImportError:  # pragma: no cover
     pass